Made deleting invoice entries possible
Now user can delete invoice entry by selecting it and clicking right button. It is also worth mentioning that in this commit a new AutoComobobox component was developed, but some features of it are still buggy. It might work perfectly given some more time, but due to the fact that Productor requires a major GUI rewamp which will probably require moving away from tkinter library, the work was given up.